GG - Fixed account deletion & signOut

+++ ProjetoP3/Core/Authentication/ViewModel/AuthViewModel.swift
@@ -69,14 +69,10 @@ class AuthViewModel: ObservableObject {
     }
     
     //selfUser: the current logged in user decided do delete is own account
-    func deleteAccount(selfUser: Bool, user: FirebaseAuth.User) {
-        user.delete() // deletes user from the auth table
-        Firestore.firestore().collection("users").document(user.uid).delete() // deletes user metadate from users collection
-        
-        if selfUser {
-            clearSessionData()
-        }
-
+    func deleteAccount() async throws {
+        guard let user = Auth.auth().currentUser else { return } // deletes user from the auth table
+        try await Firestore.firestore().collection("users").document(user.uid).delete() // deletes user metadate from users collection
+        clearSessionData()
     }
     
     func fetchUser() async throws {
